+++ CHANGES.md
@@ -3,6 +3,8 @@
 - add `set_all` and `clear_all`
 - add basic testing
 - add endianness in sexp conversion
+- add `is_empty` and `is_full`
+- add set operations
 
 ## 0.0.2
 - Initial release